refactoring: making somethings better

+++ src/main/java/com/eduardocarlos/productivityApp/models/dtos/AddHoursDTO.java
@@ -1,4 +1,7 @@
 package com.eduardocarlos.productivityApp.models.dtos;
 
-public record AddHoursDTO(Integer month, Integer year, Long hours) {
+import java.math.BigDecimal;
+
+public record AddHoursDTO(Integer month, Integer year, BigDecimal hours) {
+
 }
